feat: rework deadlock handling
The current practice of filtering specific access masks out that
can cause deadlocks in NtQueryObject also causes other handles,
e.g. most named pipes, to be ignored.
To improve on this, use a native thread instead that can be killed
in case of a deadlock. Our "main" go routine can communicate with
the go function running on the native thread via a pair of pipes.