feat: Use fully native thread
Previously, we ran a go function over a native thread that was
killed. However, the Golang scheduler remained under the
(incorrect) assumption that the go function was still running and
simply hanging, meaning that the g internally was never freed and
was e.g. printed when a stack trace was created.
To circumvent this, we now use CGO to run NtQueryObject in a fully
native thread. The native thread uses events to communicate with
the Go threads and is simply a thin wrapper around NtQueryObject.